Add images of agitator
The inclusion of the base plate is not explicitly defined in the readme or in any previous images. The pack is sized so that this can be made out of an old CD or DVD. I still have cardstock in mine because it hasn't failed yet, but CDs or DVDs would work better (last longer).